Refactor engine and tidy types - preparing for book support
Refactored the engine to use a new struct to store context, removed globals like the node count `nodes` and transposition table `table`.

Now, these globals live alongside the time limit and (new) depth limit in struct `engine.SearchContext`

Also tidied types in general, removed type aliasing and unnecessary conversions. Still some awkwardness in the engine (must convert `ply` from `uint8` to `int16` to subtract/add from/to `WIN_SCORE`), but overall it is much better.

+++ src/engine/eval.go
@@ -4,7 +4,7 @@ import (
 	"github.com/eli-rich/goc4/src/board"
 )
 
-func Eval(b *board.Board) int {
+func Eval(b *board.Board) int16 {
 
 	pboard := b.Bitboards[b.Turn]
 	oboard := b.Bitboards[b.Turn^1]
@@ -12,7 +12,7 @@ func Eval(b *board.Board) int {
 	playerRemain := board.WinsRemaining(oboard)
 	oppRemain := board.WinsRemaining(pboard)
 
-	return int(playerRemain - oppRemain)
+	return int16(playerRemain - oppRemain)
 }
 
 func CheckWinner(b *board.Board) int8 {
